Buttons with Functions
Home screen buttons are clickable now, which navigates to product's scene. This was able to do by writting script named OpenProductScript which can be found on Scripts folder. Also, a new Scene is created, to hold the products. In this scene you also can find a Home button, which navigates you to Home screen (Start). This was able to do by writting a script named GoBackHomeScript which can be found under Scripts Folder.